pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.7",
    "datasets>=4.4.2",
    "faiss-cpu>=1.13.1",
    "ipykernel>=7.1.0",
    "ipywidgets>=8.1.8",
    "langchain>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.4",
    "langchain-groq>=1.1.1",
    "langchain-text-splitters>=1.1.0",
    "langgraph>=1.0.5",
    "matplotlib>=3.10.8",
    "notebook>=7.5.1",
    "pandas>=2.3.3",
    "pyvis>=0.3.2",
    "rank-bm25>=0.2.2",
    "rapidfuzz>=3.14.3",
    "sentence-transformers>=5.2.0",
    "tavily-python>=0.7.17",
    "tqdm>=4.67.1",
    "wikipedia>=1.4.0",
    "wikipedia-api>=0.8.1",
]
